remove cds-dk installation from build script of db (#184)
This PR removes the @sap/cds-dk installation from the build script in
db/package.json. Now, the build script relies on a previous local
installation of the cds-dk, e.g. done with a maven build. Or a globally
installed cds-dk version is used.
The mbt build works fine, because first the srv module is built, then
the db module.